Added .gitignore + Added Initial Maps
Added .gitignore to exclude basic assets (Both Source related and not), as well as compiled BSP files, as they can get large in scope.

Added Initial Maps to the repository (Their VMF and VMX files). These may not be the newest versions of these maps, but they are the newest I could find from my backups